backend/requirements.txt
# Hyglow API — pin major versions for reproducible deploys
fastapi[standard]==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
alembic==1.14.0
psycopg[binary]==3.2.3
pydantic-settings==2.6.1
bcrypt==4.2.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.17
email-validator==2.2.0
